Move `FrequencyAwareFeature` into `hpotk.model`.
