+++ src/hpotk/annotations/_api.py
@@ -2,7 +2,7 @@ import abc
 import typing
 import warnings
 
-from hpotk.model import Identified, ObservableFeature, Versioned
+from hpotk.model import Identified, FrequencyAwareFeature, Versioned
 from hpotk.model import TermId
 
 
@@ -11,75 +11,6 @@ from hpotk.model import TermId
 # An example application includes phenotypic features (annotations), diseases or samples (annotated items),
 # and OMIM corpus (annotated item containers).
 # #####################################################################################################################
-
-
-class FrequencyAwareFeature(ObservableFeature, metaclass=abc.ABCMeta):
-    """
-    `FrequencyAwareFeature` entities describe the frequency of a feature in one or more annotated items.
-
-    The simplest case is presence or absence of the feature in a single item, for instance the presence or absence
-    of a phenotypic feature, such as hypertension, in a study subject. Another use case is representation
-    of the feature frequency in a collection of items, such as presence of a phenotypic feature in a cohort.
-
-    The absolute counts are stored in the `numerator` and `denominator` attributes.
-
-    **IMPORTANT**: the implementor must check the following:
-     - the `numerator` must be a non-negative `int`
-     - the `denominator` must be a positive `int`
-
-    Use the convenience static method `FrequencyAwareFeature.check_numerator_and_denominator` to check the properties.
-    """
-
-    @property
-    @abc.abstractmethod
-    def numerator(self) -> int:
-        """
-        :return: a non-negative `int` representing the count of annotated items where the annotation was present.
-        """
-        pass
-
-    @property
-    @abc.abstractmethod
-    def denominator(self) -> int:
-        """
-        :return: a positive `int` representing the total count of annotated items investigated
-        for presence/absence of an annotation.
-        """
-        pass
-
-    def frequency(self) -> float:
-        """
-
-        :return: a `float` in range :math:`[0, 1]` representing the ratio of the annotation in the annotated item(s).
-        """
-        return self.numerator / self.denominator
-
-    @property
-    def is_present(self) -> bool:
-        """
-        :return: `True` if the annotation was observed in one or more items.
-        """
-        return self.numerator != 0
-
-    @property
-    def is_excluded(self) -> bool:
-        """
-        :return: `True` if the annotation was observed in none of the annotated item(s), and therefore, excluded.
-        """
-        return self.numerator == 0
-
-    @staticmethod
-    def check_numerator_and_denominator(numerator: int, denominator: int) -> None:
-        """
-        Check if the `numerator` and `denominator` satisfy the requirements described in :class:`FrequencyAwareFeature`.
-
-        :return: `None` if the check passes or raises a `ValueError` if the `numerator` or `denominator` contain
-          invalid values.
-        """
-        if not isinstance(numerator, int) or numerator < 0:
-            raise ValueError(f'Numerator {numerator} must be a non-negative `int`')
-        if not isinstance(denominator, int) or denominator <= 0:
-            raise ValueError(f'Denominator {denominator} must be a positive `int`')
 
 
 class AnnotationBase(Identified, FrequencyAwareFeature, metaclass=abc.ABCMeta):
